feat: initial works to adapt agent to linux
