+++ .gitignore
@@ -3,3 +3,4 @@ openuem.json
 *.exe
 *-log.txt
 *.log
+certificates
