+++ go.mod
@@ -19,7 +19,7 @@ require (
 	github.com/labstack/echo/v4 v4.12.0
 	github.com/nats-io/nats.go v1.39.1
 	github.com/open-uem/nats v0.0.0-20250324100902-e34b427bb2cf
-	github.com/open-uem/utils v0.0.0-20250105115248-b0599a18d2b5
+	github.com/open-uem/utils v0.0.0-20250214102216-df5a38667d0a
 	github.com/open-uem/wingetcfg v0.0.0-20250317130843-80896d796f35
 	github.com/pkg/sftp v1.13.6
 	golang.org/x/crypto v0.36.0
@@ -51,8 +51,10 @@ require (
 	github.com/nats-io/nuid v1.0.1 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/robfig/cron/v3 v3.0.1 // indirect
+	github.com/safchain/ethtool v0.5.10 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasttemplate v1.2.2 // indirect
+	github.com/zcalusic/sysinfo v1.1.3 // indirect
 	go.opencensus.io v0.24.0 // indirect
 	golang.org/x/exp v0.0.0-20240613232115-7f521ea00fb8 // indirect
 	golang.org/x/text v0.23.0 // indirect
