+++ internal/agent/agent_linux.go
@@ -1,0 +1,9 @@
+//go:build linux
+
+package agent
+
+import "errors"
+
+func (a *Agent) StartVNCSubscribe() error {
+	return errors.New("not implemented in Linux, yet")
+}
