+++ internal/commands/report/antivirus_linux.go
@@ -1,0 +1,12 @@
+//go:build linux
+
+package report
+
+import (
+	"errors"
+)
+
+// TODO LINUX
+func (r *Report) getAntivirusInfo(debug bool) error {
+	return errors.New("not implemented in Linux, yet")
+}
