+++ internal/commands/report/cancel_windows.go
@@ -1,3 +1,5 @@
+//go:build windows
+
 package report
 
 import (
