+++ internal/commands/report/logical_disks_linux.go
@@ -1,0 +1,3 @@
+//go:build linux
+
+package report
