+++ internal/commands/report/monitor_linux.go
@@ -1,0 +1,10 @@
+//go:build linux
+
+package report
+
+import "errors"
+
+// TODO LINUX
+func (r *Report) getMonitorsInfo(debug bool) error {
+	return errors.New("not implemented in Linux, yet")
+}
