+++ internal/commands/report/share.go
@@ -1,25 +1,8 @@
 package report
 
 import (
-	"context"
 	"fmt"
-	"log"
 )
-
-func (r *Report) getSharesInfo(debug bool) error {
-	if debug {
-		log.Println("[DEBUG]: shares info has been requested")
-	}
-
-	err := r.getSharesFromWMI()
-	if err != nil {
-		log.Printf("[ERROR]: could not get shares information from WMI Win32_Share: %v", err)
-		return err
-	} else {
-		log.Printf("[INFO]: shares information has been retrieved from WMI Win32_Share")
-	}
-	return nil
-}
 
 func (r *Report) logShares() {
 	fmt.Printf("\n** 📤 Shares ********************************************************************************************************\n")
@@ -35,16 +18,4 @@ func (r *Report) logShares() {
 	} else {
 		fmt.Printf("%-40s\n", "No shares found")
 	}
-}
-
-func (r *Report) getSharesFromWMI() error {
-	namespace := `root\cimv2`
-	qShares := "SELECT Name, Path, Description FROM Win32_Share"
-
-	ctx := context.Background()
-	err := WMIQueryWithContext(ctx, qShares, &r.Shares, namespace)
-	if err != nil {
-		return err
-	}
-	return nil
 }
