+++ internal/commands/vnc/runtime_windows.go
@@ -1,3 +1,5 @@
+//go:build windows
+
 package vnc
 
 import (
