+++ internal/logger/logger.go
@@ -1,0 +1,11 @@
+package logger
+
+import "os"
+
+type OpenUEMLogger struct {
+	LogFile *os.File
+}
+
+func (l *OpenUEMLogger) Close() {
+	l.LogFile.Close()
+}
